+++ src/test/java/uk/gov/hmcts/ethos/replacement/docmosis/service/TseRespondentReplyServiceTest.java
@@ -2,22 +2,34 @@ package uk.gov.hmcts.ethos.replacement.docmosis.service;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.MockedStatic;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
+import uk.gov.hmcts.ecm.common.helpers.UtilHelper;
 import uk.gov.hmcts.ecm.common.idam.models.UserDetails;
+import uk.gov.hmcts.et.common.model.bulk.types.DynamicValueType;
 import uk.gov.hmcts.et.common.model.ccd.CaseData;
 import uk.gov.hmcts.et.common.model.ccd.CaseDetails;
+import uk.gov.hmcts.et.common.model.ccd.items.GenericTseApplicationType;
+import uk.gov.hmcts.et.common.model.ccd.items.GenericTseApplicationTypeItem;
+import uk.gov.hmcts.et.common.model.ccd.types.TseRespondType;
 import uk.gov.hmcts.ethos.replacement.docmosis.config.NotificationProperties;
 import uk.gov.hmcts.ethos.replacement.docmosis.helpers.HelperTest;
 import uk.gov.hmcts.ethos.replacement.docmosis.helpers.TseHelper;
 import uk.gov.hmcts.ethos.utils.CaseDataBuilder;
+import uk.gov.hmcts.ethos.utils.TseApplicationBuilder;
 
-import java.io.IOException;
+import java.time.LocalDate;
 import java.util.Collections;
+import java.util.List;
+import java.util.UUID;
 
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.core.Is.is;
+import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.ArgumentMatchers.eq;
@@ -26,8 +38,12 @@ import static org.mockito.Mockito.mockStatic;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
+import static uk.gov.hmcts.ecm.common.model.helper.Constants.CLAIMANT_TITLE;
 import static uk.gov.hmcts.ecm.common.model.helper.Constants.NO;
+import static uk.gov.hmcts.ecm.common.model.helper.Constants.OPEN_STATE;
+import static uk.gov.hmcts.ecm.common.model.helper.Constants.RESPONDENT_TITLE;
 import static uk.gov.hmcts.ecm.common.model.helper.Constants.YES;
+import static uk.gov.hmcts.ethos.replacement.docmosis.utils.DocumentTypeItemUtil.createSupportingMaterial;
 
 @ExtendWith(SpringExtension.class)
 class TseRespondentReplyServiceTest {
@@ -42,7 +58,7 @@ class TseRespondentReplyServiceTest {
     private NotificationProperties notificationProperties;
     private UserDetails userDetails;
     private CaseData caseData;
-    private MockedStatic mockStatic;
+    private MockedStatic<TseHelper> mockStatic;
 
     @BeforeEach
     void setUp() throws Exception {
@@ -65,6 +81,16 @@ class TseRespondentReplyServiceTest {
             .withRespondent("respondent", YES, "01-Jan-2003", false)
             .build();
 
+        GenericTseApplicationType build = TseApplicationBuilder.builder().withApplicant(CLAIMANT_TITLE)
+            .withDate("13 December 2022").withDue("20 December 2022").withType("Withdraw my claim")
+            .withCopyToOtherPartyYesOrNo(YES).withDetails("Text").withNumber("1")
+            .withResponsesCount("0").withStatus(OPEN_STATE).build();
+
+        GenericTseApplicationTypeItem genericTseApplicationTypeItem = new GenericTseApplicationTypeItem();
+        genericTseApplicationTypeItem.setId(UUID.randomUUID().toString());
+        genericTseApplicationTypeItem.setValue(build);
+        caseData.setGenericTseApplicationCollection(List.of(genericTseApplicationTypeItem));
+
         caseData.setClaimant("Claimant LastName");
     }
 
@@ -73,8 +99,47 @@ class TseRespondentReplyServiceTest {
         mockStatic.close();
     }
 
+    @Nested
+    class SaveReplyToApplication {
+        @Test
+        void whenNoApplications_doesNothing() {
+            caseData.setGenericTseApplicationCollection(null);
+            tseRespondentReplyService.saveReplyToApplication(caseData);
+            assertNull(caseData.getGenericTseApplicationCollection());
+        }
+
+        @Test
+        void savesReplyCorrectly() {
+            caseData.setTseRespondSelectApplication(TseHelper.populateRespondentSelectApplication(caseData));
+            caseData.getTseRespondSelectApplication().setValue(DynamicValueType.create("1", ""));
+
+            caseData.setTseResponseText("ResponseText");
+            caseData.setTseResponseSupportingMaterial(createSupportingMaterial());
+
+            caseData.setTseResponseHasSupportingMaterial(YES);
+            caseData.setTseResponseCopyToOtherParty(NO);
+            caseData.setTseResponseCopyNoGiveDetails("It's a secret");
+
+            tseRespondentReplyService.saveReplyToApplication(caseData);
+
+            TseRespondType replyType = caseData.getGenericTseApplicationCollection().get(0)
+                .getValue().getRespondCollection().get(0).getValue();
+
+            String dateNow = UtilHelper.formatCurrentDate(LocalDate.now());
+
+            assertThat(replyType.getDate(), is(dateNow));
+            assertThat(replyType.getResponse(), is("ResponseText"));
+            assertThat(replyType.getCopyNoGiveDetails(), is("It's a secret"));
+            assertThat(replyType.getHasSupportingMaterial(), is(YES));
+            assertThat(replyType.getCopyToOtherParty(), is(NO));
+            assertThat(replyType.getFrom(), is(RESPONDENT_TITLE));
+            assertThat(replyType.getSupportingMaterial().get(0).getValue().getUploadedDocument().getDocumentFilename(),
+                is("image.png"));
+        }
+    }
+
     @Test
-    void sendAcknowledgementAndClaimantEmail_rule92Yes() throws IOException {
+    void sendAcknowledgementAndClaimantEmail_rule92Yes() {
         caseData.setTseResponseCopyToOtherParty(YES);
 
         CaseDetails caseDetails = new CaseDetails();
@@ -89,7 +154,7 @@ class TseRespondentReplyServiceTest {
     }
 
     @Test
-    void sendAcknowledgementAndClaimantEmail_rule92No() throws IOException {
+    void sendAcknowledgementAndClaimantEmail_rule92No() {
         caseData.setTseResponseCopyToOtherParty(NO);
 
         CaseDetails caseDetails = new CaseDetails();
@@ -100,5 +165,25 @@ class TseRespondentReplyServiceTest {
 
         verify(emailService, times(1)).sendEmail(any(), any(), any());
         verify(emailService).sendEmail(any(), eq(userDetails.getEmail()), any());
+    }
+
+    @Test
+    void resetReplyToApplicationPage_resetsData() {
+        caseData.setTseResponseCopyToOtherParty(YES);
+        caseData.setTseResponseCopyNoGiveDetails(YES);
+        caseData.setTseResponseText(YES);
+        caseData.setTseResponseIntro(YES);
+        caseData.setTseResponseTable(YES);
+        caseData.setTseResponseHasSupportingMaterial(YES);
+        caseData.setTseResponseSupportingMaterial(createSupportingMaterial());
+        tseRespondentReplyService.resetReplyToApplicationPage(caseData);
+
+        assertNull(caseData.getTseResponseText());
+        assertNull(caseData.getTseResponseIntro());
+        assertNull(caseData.getTseResponseTable());
+        assertNull(caseData.getTseResponseHasSupportingMaterial());
+        assertNull(caseData.getTseResponseSupportingMaterial());
+        assertNull(caseData.getTseResponseCopyToOtherParty());
+        assertNull(caseData.getTseResponseCopyNoGiveDetails());
     }
 }
